Hacky handler for array validation where the array needs fixed items
There are certain properties like HTTP verbs where you can only have
particular combinations of values despite the property being an array.

e.g. ['HEAD', 'GET'] but not just ['GET']

This is a rather hacky solution to it, because prior to this setting
a Schema to Array:true destructured the array and tested each item
individually with the ValidateFunc -- this doesn't work with this style
of Array as you need to validate the entire thing; not specifying
Array:true would fail as the array then is tested to see if it's a
String.

It would be nice to find a way to reconcile this.