Separate Schema and Template
The validation was spread across Schema (representing the AWS resource
definitions) and Template (the parsed JSON), and this was leading to
some confusion around where behaviour should belong. Being as
ultimately it's the PropertyType's which validate themselves (and them
being a part of the Schema) it made sense to move ALL validation to the
Schema and make the Template structs entirely dumb.

This is a large refactoring and has left behind a few hairy bits;
however, it should lead to a more obvious design.