Move Array quality into the Type
Prior to this the Array marker was a part of the Schema which the Type
was a part; however, this meant that in places where we just compare
the types the Array quality wasn't being considered.

With this change Array is now inherently part of the type
(e.g. String != List<String>)